Refactor class selection popover e2e tests
