+++ cypress/integration/class-selection-popover.common.ts
@@ -1,78 +1,10 @@
-import { gql } from "@apollo/client";
-
-import { LabelCreateInput, LabelType } from "../../typescript/graphql-types";
-import { distantDatabaseClient as client } from "../../typescript/web/src/connectors/apollo-client/client";
-
-const createDataset = async (name: string) => {
-  const mutationResult = await client.mutate({
-    mutation: gql`
-      mutation createDataset($name: String) {
-        createDataset(
-          data: { name: $name, workspaceSlug: "cypress-test-workspace" }
-        ) {
-          id
-          slug
-        }
-      }
-    `,
-    variables: {
-      name,
-    },
-  });
-
-  const {
-    data: {
-      createDataset: { id, slug },
-    },
-  } = mutationResult;
-  return { id, slug };
-};
-
-async function createImage(url: string, datasetId: string) {
-  const mutationResult = await client.mutate({
-    mutation: gql`
-      mutation createImage($url: String, $datasetId: ID!) {
-        createImage(data: { url: $url, datasetId: $datasetId }) {
-          id
-          name
-          width
-          height
-          url
-        }
-      }
-    `,
-    variables: {
-      datasetId,
-      url,
-    },
-  });
-
-  const {
-    data: { createImage: image },
-  } = mutationResult;
-
-  return image;
-}
-
-const createLabel = (data: LabelCreateInput) => {
-  return client.mutate({
-    mutation: gql`
-      mutation createLabel($data: LabelCreateInput!) {
-        createLabel(data: $data) {
-          id
-        }
-      }
-    `,
-    variables: {
-      data,
-    },
-  });
-};
+import { ApolloClient, gql, NormalizedCacheObject } from "@apollo/client";
 
 const createLabelClass = async (
   name: String,
   color = "#ffffff",
-  datasetId: string
+  datasetId: string,
+  client: ApolloClient<NormalizedCacheObject>
 ) => {
   const {
     data: {
@@ -104,55 +36,25 @@ const createLabelClass = async (
   return id;
 };
 
-describe("Class selection popover", () => {
-  let datasetId: string;
-  let datasetSlug: string;
-  let imageId: string;
-  beforeEach(() => {
-    cy.setCookie("consentedCookies", "true");
-    cy.task("performLogin").then((token) => {
-      cy.setCookie("next-auth.session-token", token as string);
-    });
-    cy.task("createWorkspaceAndDatasets").then(async () => {
-      const createResult = await createDataset("cypress test dataset");
-      datasetId = createResult.id;
-      datasetSlug = createResult.slug;
-
-      const { id } = await createImage(
-        "https://images.unsplash.com/photo-1579513141590-c597876aefbc?auto=format&fit=crop&w=882&q=80",
-        datasetId
-      );
-      imageId = id;
-
-      const labelClassId = await createLabelClass(
-        "A new class",
-        "#F87171",
-        datasetId
-      );
-      await createLabel({
-        imageId,
-        labelClassId,
-        type: LabelType.Box,
-        geometry: {
-          type: LabelType.Polygon,
-          coordinates: [
-            [
-              [0, 900],
-              [900, 900],
-              [900, 1500],
-              [0, 1500],
-              [0, 900],
-            ],
-          ],
-        },
-      });
-    });
-  });
-
+export const declareClassSelectionPopoverTests = ({
+  client,
+  workspaceSlug,
+  getDatasetId,
+  getDatasetSlug,
+  getImageId,
+}: {
+  client: ApolloClient<NormalizedCacheObject>;
+  workspaceSlug: string;
+  getDatasetId: () => string;
+  getDatasetSlug: () => string;
+  getImageId: () => string;
+}) => {
   it("right clicks on a label to change its class", () => {
+    const datasetSlug = getDatasetSlug();
+    const imageId = getImageId();
     // See https://docs.cypress.io/guides/core-concepts/conditional-testing#Welcome-wizard
     cy.visit(
-      `/cypress-test-workspace/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
+      `/${workspaceSlug}/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
     );
     cy.get('[aria-label="loading indicator"]').should("not.exist");
     cy.get('[aria-label="Selection tool"]').click();
@@ -205,11 +107,14 @@ describe("Class selection popover", () => {
   });
 
   it("uses the class selection menu to change the class of created labels", () => {
-    cy.wrap(createLabelClass("My new class", "#65A30D", datasetId));
+    const datasetId = getDatasetId();
+    const datasetSlug = getDatasetSlug();
+    const imageId = getImageId();
+    cy.wrap(createLabelClass("My new class", "#65A30D", datasetId, client));
 
     // See https://docs.cypress.io/guides/core-concepts/conditional-testing#Welcome-wizard
     cy.visit(
-      `/cypress-test-workspace/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
+      `/${workspaceSlug}/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
     );
 
     cy.wait(420);
@@ -292,10 +197,13 @@ describe("Class selection popover", () => {
   });
 
   it("uses shortcuts to change classes", () => {
-    cy.wrap(createLabelClass("My new class", "#65A30D", datasetId));
+    const datasetId = getDatasetId();
+    const datasetSlug = getDatasetSlug();
+    const imageId = getImageId();
+    cy.wrap(createLabelClass("My new class", "#65A30D", datasetId, client));
     // See https://docs.cypress.io/guides/core-concepts/conditional-testing#Welcome-wizard
     cy.visit(
-      `/cypress-test-workspace/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
+      `/${workspaceSlug}/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
     );
     cy.get('[aria-label="loading indicator"]').should("not.exist");
     cy.get('[aria-label="Selection tool"]').click();
@@ -362,8 +270,10 @@ describe("Class selection popover", () => {
   });
 
   it("should update the label classes list when a label class is deleted", () => {
+    const datasetSlug = getDatasetSlug();
+    const imageId = getImageId();
     cy.visit(
-      `/cypress-test-workspace/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
+      `/${workspaceSlug}/datasets/${datasetSlug}/images/${imageId}?modal-welcome=closed&modal-update-service-worker=update`
     );
 
     cy.wait(420);
@@ -375,7 +285,7 @@ describe("Class selection popover", () => {
     cy.contains("classes").click();
     cy.url().should(
       "contain",
-      `/cypress-test-workspace/datasets/${datasetSlug}/classes`
+      `/${workspaceSlug}/datasets/${datasetSlug}/classes`
     );
 
     cy.wait(420);
@@ -398,4 +308,4 @@ describe("Class selection popover", () => {
       .contains("A new class")
       .should("not.exist");
   });
-});
+};
